refactor: move hook to appropriate location to untie circular dependency
