+++ src/right-hand-panel/use-close-right-hand-panel-on-selection-change.js
@@ -1,11 +1,11 @@
 import { useEffect } from 'react'
-import { useRightHandPanelContext } from '../right-hand-panel/index.js'
 import {
     useAttributeOptionComboSelection,
     useDataSetId,
     useOrgUnitId,
     usePeriodId,
 } from '../shared/index.js'
+import useRightHandPanelContext from './use-right-hand-panel-context.js'
 
 export default function useCloseRightHandPanelOnSelectionChange() {
     const { hide } = useRightHandPanelContext()
